Auto stash before checking out "HEAD"
Fix bug

+++ MathBrick/MathBrick/Model/Quiz.cs
@@ -15,7 +15,7 @@ namespace MathBrick.Model
         public string teacherID { get; set; }
         public Question[] questions { get; set; }
 
-        public Dictionary<string, int> studentGrades { get; set; }
+        public Dictionary<string, double> studentGrades { get; set; }
 
         public Quiz(string subject, int level, string date, string teacherID, Question[] questions)
         {
@@ -27,7 +27,7 @@ namespace MathBrick.Model
             this.date = date;
             this.teacherID = teacherID;
             this.questions = questions;
-            this.studentGrades = new Dictionary<string, int>();
+            this.studentGrades = new Dictionary<string, double>();
         }
 
         public Quiz() { }
